issue: 3414980 Move extra api code in separate file
